+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Adding support for PCFT VarVsEff plots [!309](https://github.com/umami-hep/puma/pull/309)
 - Small improvements for vertexing plots [!305](https://github.com/umami-hep/puma/pull/305)
 - Minor change to how n_classes are calculated in confusion matrix [!306](https://github.com/umami-hep/puma/pull/306)
 - Fixing Coverage [!308](https://github.com/umami-hep/puma/pull/308)